remove those weird extra Engine classes
