+++ lib/dradis/plugins/burp.rb
@@ -6,10 +6,7 @@ module Dradis
 end
 
 require 'dradis/plugins/burp/engine'
-require 'dradis/plugins/burp/html/engine'
-require 'dradis/plugins/burp/html/field_processor'
+require 'dradis/plugins/burp/field_processor'
 require 'dradis/plugins/burp/html/importer'
 require 'dradis/plugins/burp/version'
-require 'dradis/plugins/burp/xml/engine'
-require 'dradis/plugins/burp/xml/field_processor'
 require 'dradis/plugins/burp/xml/importer'
