+++ lib/dradis/plugins/burp/engine.rb
@@ -19,12 +19,6 @@ module Dradis
             Dradis::Plugins::Burp::Xml
           ]
         end
-
-        # We define Dadis::Plugins::Burp::[Html/Xml]::Engine
-        # But we still want the plugin name to be 'burp' 
-        def self.plugin_name
-          'burp'
-        end
       end
     end
   end
