+++ lib/dradis/plugins/burp/html/importer.rb
@@ -11,6 +11,11 @@ module Dradis::Plugins::Burp
     end
 
     class Importer < Dradis::Plugins::Upload::Importer
+      def initialize(args={})
+        args[:plugin] = Dradis::Plugins::Burp
+        super(args)
+      end
+
       def import(params = {})
         logger.info { 'Parsing Burp Scanner HTML output file...' }
         @doc = Nokogiri::HTML(File.read(params[:file]))
