--- lib/dradis/plugins/burp/xml/engine.rb
@@ -1,3 +1,0 @@
-module Dradis::Plugins::Burp::Xml
-  class Engine < Dradis::Plugins::Burp::Engine; end
-end
